+++ src/main/java/com/jiyong/commerce/common/aop/AspectConfig.java
@@ -1,0 +1,11 @@
+package com.jiyong.commerce.common.aop;
+
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Aspect
+@Component
+public class AspectConfig {
+}
