+++ src/main/java/com/jiyong/commerce/item/repository/MemoryItemRepository.java
@@ -1,10 +1,8 @@
-package com.jiyong.commerce.repository;
+package com.jiyong.commerce.item.repository;
 
-import com.jiyong.commerce.domain.item.Item;
+import com.jiyong.commerce.item.domain.Item;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -14,6 +12,12 @@ public class MemoryItemRepository implements ItemRepository{
 
     private static Map<Long, Item> store = new ConcurrentHashMap<>();
     private static long sequence = 0L;
+
+    //ToDo
+    /**
+     * 성공률 99% (1% = IOException, SocketException, TimeoutException - RuntimeException의 하위, Retryable exponentially 인지 아닌지) AuthorizationException
+     * (Math.random() 0 ~ 1) * 100 < 1
+     * */
 
     @Override
     public Item insertItem(Item item) {
